Feat:Completed the adding, editing, deleting and duplicating rooms. Also fetched and add the posibility og editing a single settings field onBlur using React Forms and Query.

+++ src/services/apirooms.js
@@ -18,27 +18,51 @@ export async function deleteRoom(roomId) {
   }
 }
 
-export async function addRoom(newRoom) {
+export async function addAndEditRoom(newRoom, id = null) {
+  console.log(id, "id");
   const imageName = `${Math.random()}-${newRoom.image.name}`.replaceAll(
     "/",
     "",
   );
+  const hasSupabaseImgPath = newRoom.image?.startsWith?.(SUPABASEURL);
 
-  const roomImagePath = `${SUPABASEURL}/storage/v1/object/public/rooms_images/${imageName}`;
-  const { data, error } = await supabase
-    .from("rooms")
-    .insert([{ ...newRoom, image: roomImagePath }])
-    .select();
+  const roomImagePath = hasSupabaseImgPath
+    ? newRoom.image
+    : `${SUPABASEURL}/storage/v1/object/public/rooms_images/${imageName}`;
+
+  let query = supabase.from("rooms");
+
+  if (!id) {
+    console.log("whattttt");
+    query = query
+      .insert([{ ...newRoom, image: roomImagePath }])
+      .select()
+      .single();
+  }
+
+  if (id) {
+    console.log("upds");
+
+    query = query
+      .update({ ...newRoom, image: roomImagePath })
+      .eq("id", id)
+      .select();
+  }
+
+  const { data, error } = await query;
+
   if (error) {
     throw new Error("Oops, something went wrong while adding room");
   }
 
-  const { error: bucketError } = await supabase.storage
-    .from("rooms_images")
-    .upload(imageName, newRoom.image);
+  if (!hasSupabaseImgPath) {
+    const { error: bucketError } = await supabase.storage
+      .from("rooms_images")
+      .upload(imageName, newRoom.image);
 
-  if (bucketError) {
-    await supabase.from("rooms").delete().eq("id", data.id);
+    if (bucketError) {
+      await supabase.from("rooms").delete().eq("id", data.id);
+    }
   }
   return data;
 }
